Calculator - Should return double when dividing two numbers (RED) - Quick fix of test in `CalculatorTest.java` to check if double is returned by divide, as it will be required in `TextAnalyzerTest.java`

+++ src/test/java/CalculatorTest.java
@@ -46,7 +46,7 @@ public class CalculatorTest {
     @Test
     @DisplayName("Should divide two numbers")
     public void shouldDivideTwoNumbers() {
-        int expected = 7;
+        double expected = 7;
         Assertions.assertEquals(expected, calculator.divide(21, 3));
     }
